Major re-work
I tweaked the way that the job object is built and organized.
ultimately this way is faster because it minimizes iteration and
builds its status object more intelligently.

jobs are now referenced by job_id which must be converted to int()
before use as an index.

--debug pretty prints the entire history object now.

Next steps are to a) get the undo and redo functions mapped back out
and b) start testing out the apt / dpkg / dpkg-reconfigure items I've
built in.