Version 0.1 Just Stable Enough
Version 0.1 -- Just Stable Enough Edition

Can parse through all of the dpkg log files in
a given /var/log dir, load them up, and organize
the lines returned into a detailed ledger of dpkg
jobs. This ledger can currently be used to show
when a job occurred, how long it took, and what
exactly happened during that time.

Currently the `--list` option will only show you
the last 30 or so days of transactions. this can
be overridden by providing either `--show-last`,
which is used to provide a specific number of days
to go back, or `--show-all`, which will print in
the table all of the jobs it can find from the
dpkg logs.

You can generate a pretty transaction report using
the `--info` flag to provide `dpkg-history` with a
job_id.

The `--inspect` flag also provides the script with
a job_id, but inspect differs from info in that it
dumps the raw logfile.

Soon `dpkg-history` will be able to:

* undo and redo transactions by job id.
* search across all jobs for packages by name.